Improve function and clarify output to user
These changes improve the function so that it may work with any number of
experiments in which each experiment is a series of 100 coin flips. The function
now correctly calculates the percentage for the number of 6-streaks that occur
in each experiment. The output is clarified for the user in a much better way
by displaying the number of experiments being processed, the percentage
chance of getting a 6-streak, and the number of 6-streaks that were found.