Finish program functionality
Added the ability for the program to insert a number of blank
rows at a specific starting row, both values given by the user.
Also added the functionality for the program to save the newly
created spreadsheet file in the same directory path as the file
given by the user.

Added a try/except clause to ensure that the first two arguments
given to the program, the starting row and the number of rows to
add, can properly be converted into integer types.